integer length issues

+++ mch2022/components/mch2022_rp2040/mch2022_rp2040.cpp
@@ -30,7 +30,7 @@ void Mch2022_rp2040Component::dump_config() {
 
 void Mch2022_rp2040Component::loop() {
     if (!this->interrupt_pin_->digital_read()) {
-      uint16_t reg1 = 0; 
+      uint8_t reg1 = 0; 
       this->read_register(RP2040_REG_INPUT1, &reg1, 1);
     // uint8_t reg2 = this->read_register(RP2040_REG_INPUT1, &port, 1);
     // ESP_LOGE(TAG, "Button pressed on the rp2040 !");
